move debug fn
